+++ hack/runtime-migrator/internal/input/client.go
@@ -1,4 +1,4 @@
-package config
+package input
 
 import (
 	"fmt"
