+++ hack/runtime-migrator/internal/input/migrationparams.go
@@ -1,4 +1,4 @@
-package config
+package input
 
 import (
 	"bufio"
